Get rid of php_dbus_data_array
In all cases it was used was to pass a copy of primitive values like
bool, long (lval), double (dval) etc. Therefore, simply use a local
variable to make a copy and pass a pointer to it. Also
dbus_message_iter_append value is declared as const void * so it does
not modify contents anyway.